# cedar-artifact-server-application/src/main/java/org/metadatacenter/cedar/artifact/resources/CommandResource.java
package org.metadatacenter.cedar.artifact.resources;

import com.codahale.metrics.annotation.Timed;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.parameters.RequestBody;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.security.SecurityRequirementEntry;
import com.fasterxml.jackson.databind.JsonNode;
import org.metadatacenter.util.artifact.ArtifactDocument;
import org.metadatacenter.util.http.CedarError;
import org.metadatacenter.config.CedarConfig;
import org.metadatacenter.exception.CedarException;
import org.metadatacenter.exception.CedarProcessingException;
import org.metadatacenter.constant.HttpConstants;
import org.metadatacenter.model.CedarResourceType;
import org.metadatacenter.model.request.ResourceType;
import org.metadatacenter.model.request.ResourceTypeDetector;
import org.metadatacenter.model.validation.report.CedarValidationReport;
import org.metadatacenter.model.validation.report.ValidationReport;
import org.metadatacenter.rest.context.CedarRequestContext;
import org.metadatacenter.server.service.TemplateService;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import jakarta.ws.rs.Consumes;
import jakarta.ws.rs.POST;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.QueryParam;
import jakarta.ws.rs.core.MediaType;
import jakarta.ws.rs.core.Response;
import java.io.IOException;

import static com.google.common.base.Preconditions.checkNotNull;
import static org.metadatacenter.constant.CedarQueryParameters.QP_RESOURCE_TYPE;
import static org.metadatacenter.rest.assertion.GenericAssertions.LoggedIn;

@Path("/command")
@Produces(MediaType.APPLICATION_JSON)
@Tag(name = "Command")
@SecurityRequirement(name = "", combine = {
    @SecurityRequirementEntry(name = "api_key"), @SecurityRequirementEntry(name = "artifact_service")})
public class CommandResource extends AbstractArtifactServerResource {

  private static final Logger log = LoggerFactory.getLogger(CommandResource.class);

  private final TemplateService<String, JsonNode> templateService;

  public CommandResource(CedarConfig cedarConfig, TemplateService<String, JsonNode> templateService) {
    super(cedarConfig);
    this.templateService = checkNotNull(templateService);
  }

  /**
   * Validates an artifact given in the request body, in either serialization CEDAR speaks.
   *
   * <p>YAML is transcoded by the same path create and update use, so a client that authors in YAML can
   * ask whether its work is valid before sending it. This route was JSON only, and a YAML body reached
   * Jackson and failed to deserialize — a 500 for what is a client's business, and the one
   * write-adjacent route that did not negotiate what the write routes accept.
   *
   * <p>An instance may also be validated against a template supplied alongside it, as
   * {@code {"schema": …, "instance": …}}. That composite is a JSON convenience rather than an artifact,
   * so it has no YAML form: a YAML body is read as the artifact itself.
   */
  @POST
  @Timed
  @Path("/validate")
  @Consumes({MediaType.APPLICATION_JSON, HttpConstants.CONTENT_TYPE_APPLICATION_YAML, "application/yaml"})
  @Operation(summary = "Validate an artifact",
      description = "Validate an artifact against the CEDAR model without storing it. The body may be "
          + "JSON or YAML, selected by the Content-Type header, so a client that authors in YAML can "
          + "ask whether its work is valid before sending it. An instance may instead be validated "
          + "against a template supplied alongside it, as `{\"schema\": …, \"instance\": …}`; that "
          + "composite is a JSON convenience rather than an artifact, so it has no YAML form and a "
          + "YAML body is read as the artifact itself. The report is returned with 200 whether or not "
          + "the artifact is valid: an invalid artifact is an answer, not a failed request.")
  @RequestBody(required = true, description = "The artifact to validate, or an instance and the "
      + "template to validate it against.",
      content = {
          @Content(mediaType = MediaType.APPLICATION_JSON,
              schema = @Schema(ref = "#/components/schemas/ValidationTarget")),
          @Content(mediaType = HttpConstants.CONTENT_TYPE_APPLICATION_YAML,
              schema = @Schema(implementation = ArtifactDocument.class)),
          @Content(mediaType = "application/yaml",
              schema = @Schema(implementation = ArtifactDocument.class))
      })
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "The validation report",
          content = @Content(schema = @Schema(implementation = CedarValidationReport.class))),
      @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "The body could not be read as the named artifact type"),
      @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Unauthorized"),
      @ApiResponse(responseCode = "500", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "The validation could not be run")
  })
  public Response validateResource(
      @Parameter(description = "Artifact type to validate against: `template`, `element`, `field`, "
          + "or `instance`.", required = true)
      @QueryParam(QP_RESOURCE_TYPE) String type,
      String requestBody) throws CedarException {
    CedarRequestContext c = buildRequestContext();
    c.must(c.user()).be(LoggedIn);

    ResourceType resourceType = ResourceTypeDetector.detectType(type == null ? "" : type);
    JsonNode resourceNode = artifactRequestBody(requestBody, CedarResourceType.forValue(resourceType.getValue())).asJson();
    ValidationReport validationReport = validateResource(resourceNode, resourceType);
    return Response.ok().entity(validationReport).build();
  }

  private ValidationReport validateResource(JsonNode resource, ResourceType type) throws CedarException {
    // A switch expression over the enum, with no default: a type this cannot validate is a compile
    // error, not a runtime 500 with an error key and no status.
    return switch (type) {
      case TEMPLATE -> validateTemplate(resource);
      case ELEMENT -> validateTemplateElement(resource);
      case FIELD -> validateTemplateField(resource);
      case INSTANCE -> validateTemplateInstance(resource);
    };
  }

  private ValidationReport validateTemplateInstance(JsonNode payload) throws CedarException {
    try {
      ValidationReport validationReport = null;
      JsonNode instanceObject = getInstanceObject(payload);
      if (hasUserSpecifiedSchema(payload)) {
        validationReport = validateUsingUserSpecifiedSchema(payload, instanceObject);
      } else {
        validationReport = validateUsingInstanceSpecifiedSchema(instanceObject);
      }
      return validationReport;
    } catch (IOException e) {
      throw new CedarProcessingException(e);
    }
  }

  private boolean hasUserSpecifiedSchema(JsonNode payload) {
    return !payload.path("schema").isMissingNode() && !payload.path("schema").isNull();
  }

  private ValidationReport validateUsingUserSpecifiedSchema(JsonNode payload, JsonNode instanceObject) throws CedarException {
    JsonNode instanceSchema = payload.get("schema");
    ValidationReport schemaValidationReport = validateTemplate(instanceSchema); // validate the input schema
    if (schemaValidationReport.getValidationStatus().equals("false")) {
      return schemaValidationReport; // return schema validation report instead
    }
    return validateTemplateInstance(instanceObject, instanceSchema);
  }

  private ValidationReport validateUsingInstanceSpecifiedSchema(JsonNode instanceObject) throws IOException, CedarException {
    JsonNode instanceSchema = getSchemaSource(templateService, instanceObject);
    return validateTemplateInstance(instanceObject, instanceSchema);
  }

  private static JsonNode getInstanceObject(JsonNode payload) throws CedarException {
    JsonNode instanceObject = payload;
    if (!payload.path("instance").isMissingNode()) {
      instanceObject = payload.get("instance");
    }
    return instanceObject;
  }

}