# cedar-artifact-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "Better Metadata - Better Science\n\nInternal service: business routes require both X-CEDAR-Artifact-Service-Key and the\nend-user Authorization header. Public callers use resource; a user API key alone is insufficient.\n\nThe artifact server stores templates, template elements, template fields, and template\ninstances, and validates every one of them against the CEDAR model before storing it.\n\nArtifacts use strong ETags for optimistic concurrency. Read the artifact and send the returned\nETag in `If-Match` when replacing or deleting it. A missing required precondition returns 428;\na stale or malformed validator returns 412. `If-Match: *` accepts whichever revision currently\nexists and therefore deliberately gives up lost-update protection. Representation suffixes such\nas `\"7-yaml\"` and `\"7-yaml-compact-v2\"` identify different bytes of the same stored revision and\nare valid write preconditions.\n\nArtifacts are read and written as JSON or YAML, selected by the Accept and Content-Type\nheaders. The compact YAML form is a lossy read-time convenience: it can be read but not stored.\n",
    "title" : "CEDAR Artifact Server API",
    "version" : "2.9.11-SNAPSHOT"
  },
  "paths" : {
...
    "/command/validate" : {
      "post" : {
        "description" : "Validate an artifact against the CEDAR model without storing it. The body may be JSON or YAML, selected by the Content-Type header, so a client that authors in YAML can ask whether its work is valid before sending it. An instance may instead be validated against a template supplied alongside it, as `{\"schema\": …, \"instance\": …}`; that composite is a JSON convenience rather than an artifact, so it has no YAML form and a YAML body is read as the artifact itself. The report is returned with 200 whether or not the artifact is valid: an invalid artifact is an answer, not a failed request.",
        "operationId" : "validateResource",
        "parameters" : [ {
          "description" : "Artifact type to validate against: `template`, `element`, `field`, or `instance`.",
          "in" : "query",
          "name" : "resource_type",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ValidationTarget"
              }
            },
            "application/x-yaml" : {
              "schema" : {
                "$ref" : "#/components/schemas/ArtifactDocument"
              }
            },
            "application/yaml" : {
              "schema" : {
                "$ref" : "#/components/schemas/ArtifactDocument"
              }
            }
          },
          "description" : "The artifact to validate, or an instance and the template to validate it against.",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidationReport"
                }
              }
            },
            "description" : "The validation report"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The body could not be read as the named artifact type"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The validation could not be run"
          }
        },
        "security" : [ {
          "api_key" : [ ],
          "artifact_service" : [ ]
        } ],
        "summary" : "Validate an artifact",
        "tags" : [ "Command" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ArtifactDocument" : {
        "type" : "object",
        "additionalProperties" : true,
        "description" : "An open CEDAR JSON-LD artifact: the identity and provenance keys that templates, elements, fields and instances all carry. Kind-specific keys are described by SchemaArtifactDocument and InstanceArtifactDocument; the authoritative definition of each kind is the CEDAR model's JSON Schema.",
        "externalDocs" : {
          "description" : "The CEDAR template model",
          "url" : "https://more.metadatacenter.org/tools-training/outreach/cedar-template-model"
        },
        "properties" : {
          "@context" : {
            "type" : "object",
            "additionalProperties" : true,
            "description" : "JSON-LD context for the keys the artifact uses."
          },
          "@id" : {
            "type" : "string",
            "format" : "uri",
            "description" : "The artifact's IRI. The server mints it on a POST, so a body sent to POST carries none; a body sent to PUT carries the IRI of the path."
          },
          "oslc:modifiedBy" : {
            "type" : "string",
            "format" : "uri",
            "description" : "IRI of the user who last changed it."
          },
          "pav:createdBy" : {
            "type" : "string",
            "format" : "uri",
            "description" : "IRI of the user who created it."
          },
          "pav:createdOn" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "When the artifact was created."
          },
          "pav:lastUpdatedOn" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "When the artifact was last changed."
          },
          "schema:description" : {
            "type" : "string",
            "description" : "Description shown with the name."
          },
          "schema:name" : {
            "type" : "string",
            "description" : "Display name."
          }
        }
      },
      "CedarError" : {
        "type" : "object",
        "additionalProperties" : true,
        "description" : "A CEDAR error response. Diagnostic fields are populated only when they are relevant to the failure.",
        "properties" : {
          "entities" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named domain entities associated with exception-mapped errors.",
              "nullable" : true
            },
            "description" : "Named domain entities associated with exception-mapped errors.",
            "nullable" : true
          },
          "errorId" : {
            "type" : "string",
            "format" : "uuid",
            "description" : "Correlation identifier for an internal error recorded in server logs.",
            "nullable" : true
          },
          "errorKey" : {
            "type" : "string",
            "description" : "Stable machine-readable error identifier.",
            "enum" : [ "templateElementNotCreated", "templateElementNotFound", "templateElementNotDeleted", "templateElementNotUpdated", "templateElementsNotListed", "templateFieldNotCreated", "templateFieldNotFound", "templateFieldNotDeleted", "templateFieldNotUpdated", "templateFieldsNotListed", "templateNotCreated", "templateNotFound", "templateNotDeleted", "templateNotUpdated", "templatesNotListed", "templateInstanceNotCreated", "templateInstanceNotFound", "templateInstanceNotDeleted", "templateInstanceNotUpdated", "templateInstancesNotListed", "noReadAccessToResource", "noWriteAccessToResource", "noReadAccessToFolder", "noWriteAccessToFolder", "noReadAccessToTemplate", "noWriteAccessToTemplate", "noReadAccessToTemplateElement", "noWriteAccessToTemplateElement", "noReadAccessToTemplateField", "noWriteAccessToTemplateField", "noReadAccessToTemplateInstance", "noWriteAccessToTemplateInstance", "noReadAccessToArtifact", "noWriteAccessToArtifact", "noReadAccessToCategory", "noWriteAccessToCategory", "folderNotFound", "artifactNotFound", "artifactPreconditionRequired", "artifactHasMovedOn", "verbatimWriteRefused", "nodeNotFound", "unknownResourceType", "missingParameter", "invalidInput", "sourceFolderNotFound", "sourceResourceNotFound", "targetFolderNotFound", "nodeNotMoved", "groupNotFound", "userNotFound", "uniqueConstraintCollision", "invalidData", "notAuthorized", "permissionMissing", "tokenInvalid", "userInfoLoadByTokenFailed", "userInfoLoadByApiKeyFailed", "cedarUserNotFound", "authorizationNotFound", "permissionNotOwned", "authorizationTypeUnknown", "apiKeyNotFound", "tokenMissing", "tokenExpired", "parentFolderNotSpecified", "parentFolderSpecifiedTwice", "pathNotNormalized", "parentFolderNotFound", "updateInvalidFolderName", "createInvalidFolderName", "nodeAlreadyPresent", "folderNotCreated", "missingNameAndDescription", "missingData", "folderNotDeleted", "artifactNotDeleted", "resourceNotCreated", "invalidResourceType", "invalidArtifactType", "readOtherProfileForbidden", "updateOtherProfileForbidden", "folderCanNotBeDeleted", "folderCanNotBeChanged", "groupAlreadyPresent", "groupCanBeModifiedOnlyByGroupAdmin", "groupCanBeDeletedOnlyByGroupAdmin", "groupMembersCanBeReadOnlyByGroupAdmin", "groupRequiresAdministrator", "groupUsersNotUpdated", "specialGroupCanNotBeDeleted", "folderPermissionsCanNotBeChanged", "unknownInstanceOutputFormat", "folderCopyNotAllowed", "methodNotImplemented", "upstreamServerError", "nothingToDo", "parentCategoryNotFound", "categoryAlreadyPresent", "categoryNotFound", "categoryCanNotBeDeleted", "rootCategoryCanNotBeDeleted", "unableToAttachCategory", "unableToDetachCategory", "noCategoriesWereAttached", "malformedJsonRequestBody", "malformedSearchTerm", "publishedArtifactCanNotBeChanged", "publishedArtifactCanNotBeDeleted", "versioningOnlyOnLatest", "versioningOnlyByOwner", "nonVersionedArtifactType", "createDraftOnlyFromPublished", "publishOnlyDraft", "draftNotCreated", "contentNotValid", "doiNotSupportedByResourceType", "doiCanNotBeSetForEmptyAtId", "doiCanNotBeAltered", "doiCanNotBeSet", "doiAlreadyExists", "dataCiteDOIDisabled", "resourceNotFound", "internalError" ],
            "nullable" : true
          },
          "errorMessage" : {
            "type" : "string",
            "description" : "Alias of message retained for compatibility.",
            "nullable" : true
          },
          "errorReasonKey" : {
            "type" : "string",
            "description" : "More specific machine-readable reason for the error.",
            "enum" : [ "nonEmptyFolder", "nonEmptyCategory", "userHomeFolder", "systemFolder", "templateReferencedInInstances", "validationError", "continuationExpired" ],
            "nullable" : true
          },
          "errorType" : {
            "type" : "string",
            "description" : "Broad category of the error.",
            "enum" : [ "notFound", "invalidArgument", "authorization", "permission", "server", "validationError" ],
            "nullable" : true
          },
          "message" : {
            "type" : "string",
            "description" : "Human-readable error message.",
            "nullable" : true
          },
          "objects" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named structured values associated with the error."
            },
            "description" : "Named structured values associated with the error."
          },
          "operation" : {
            "type" : "object",
            "description" : "Operation that failed. Its fields depend on the operation type.",
            "nullable" : true
          },
          "parameters" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named scalar values associated with the error."
            },
            "description" : "Named scalar values associated with the error."
          },
          "status" : {
            "type" : "string",
            "description" : "Symbolic HTTP response status.",
            "enum" : [ "BAD_REQUEST", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "METHOD_NOT_ALLOWED", "NOT_ACCEPTABLE", "CONFLICT", "PRECONDITION_FAILED", "UNSUPPORTED_MEDIA_TYPE", "UNPROCESSABLE_ENTITY", "PRECONDITION_REQUIRED", "INTERNAL_SERVER_ERROR", "NOT_IMPLEMENTED", "BAD_GATEWAY", "SERVICE_UNAVAILABLE", "HTTP_VERSION_NOT_SUPPORTED" ]
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Numeric HTTP response status.",
            "maximum" : 599,
            "minimum" : 400
          },
          "suggestedAction" : {
            "type" : "string",
            "description" : "Recovery action suggested to the client.",
            "enum" : [ "none", "requestRole", "logout", "logoutImmediately", "provideAuthorizationHeader", "refreshToken" ],
            "nullable" : true
          }
        },
        "required" : [ "status", "statusCode" ]
      },
      "ErrorItem" : {
        "type" : "object",
        "description" : "Everything that makes the artifact invalid, in the order found.",
        "properties" : {
          "additionalInfo" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "location" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "InstanceWithSchema" : {
        "type" : "object",
        "additionalProperties" : false,
        "description" : "An instance and the template it is validated against.",
        "properties" : {
          "instance" : {
            "allOf" : [ {
              "$ref" : "#/components/schemas/ArtifactDocument"
            } ],
            "description" : "The instance to validate."
          },
          "schema" : {
            "allOf" : [ {
              "$ref" : "#/components/schemas/ArtifactDocument"
            } ],
            "description" : "The template to validate the instance against. It is validated first, and a template\nthat does not itself pass is reported in place of the instance. Omitting it validates\nthe instance against the stored template it names.\n"
          }
        },
        "required" : [ "instance" ]
      },
...
      "ValidationReport" : {
        "type" : "object",
        "description" : "What validating an artifact against the CEDAR model found. Returned with 200 whether or not the artifact is valid: an invalid artifact is an answer, not a failed request. The artifact server produces it, and the resource server hands the same document back unchanged.",
        "properties" : {
          "errors" : {
            "type" : "array",
            "description" : "Everything that makes the artifact invalid, in the order found.",
            "items" : {
              "$ref" : "#/components/schemas/ErrorItem"
            },
            "uniqueItems" : true
          },
          "validates" : {
            "type" : "string",
            "description" : "Whether the artifact validates, as a string. One error makes it false; warnings alone still validate.",
            "enum" : [ "true", "false" ]
          },
          "warnings" : {
            "type" : "array",
            "description" : "Everything the artifact should change but need not, in the order found.",
            "items" : {
              "$ref" : "#/components/schemas/WarningItem"
            },
            "uniqueItems" : true
          }
        },
        "required" : [ "errors", "validates", "warnings" ]
      },
      "ValidationTarget" : {
        "description" : "What POST /command/validate accepts as JSON: an artifact, or an instance together with the\ntemplate to validate it against. The composite form is a JSON convenience rather than an\nartifact, so it has no YAML equivalent.\n",
        "oneOf" : [ {
          "$ref" : "#/components/schemas/ArtifactDocument"
        }, {
          "$ref" : "#/components/schemas/InstanceWithSchema"
        } ]
      },
      "WarningItem" : {
        "type" : "object",
        "description" : "Everything the artifact should change but need not, in the order found.",
        "properties" : {
          "additionalInfo" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "location" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      }
    },
    "securitySchemes" : {
      "api_key" : {
        "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      },
      "artifact_service" : {
        "description" : "Private installation credential for trusted artifact callers; required alongside user authorization.",
        "in" : "header",
        "name" : "X-CEDAR-Artifact-Service-Key",
        "type" : "apiKey"
      }
    }
  }
}